Update sprite2 serialize Increase demo version Fix code not using P_GetSkinSpritedef Fix code not using correct data type for player sprite

+++ src/f_finale.c
@@ -3527,7 +3527,7 @@ void F_TitleDemoTicker(void)
 // ==========
 
 static skin_t *contskins[2];
-static UINT8 cont_spr2[2][6];
+static UINT16 cont_spr2[2][6];
 static UINT8 *contcolormaps[2];
 
 void F_StartContinue(void)
